requirements.txt
fastapi
discord.py
bardapi
asyncio
youtube-search-python
loguru
pillow